Refactor UGN-234 split validation table into separate files
